Unique Pickup Item Scrapped
Couldn't get this to work as it seems like the quest system itself doesn't play nice with the pickup item script causing null references.  The pickup item script wouldn't trigger the quest bool to false, preventing it from ever resetting.  A separate level loader would probably be best, but not looking into this.  Will now progress with tutorial and forget this bullshit.